+++ pages/comment.vue
@@ -1,0 +1,9 @@
+<template>
+  <div></div>
+</template>
+
+<script>
+export default {};
+</script>
+
+<style></style>
